requirements.txt
# Core dependencies for Consciousness Control Center
asyncio
httpx==0.25.2
requests==2.31.0
psutil==5.9.6
aiofiles==23.2.1
fastapi==0.110.0
uvicorn[standard]==0.29.0

# RAG and AI dependencies
sentence-transformers==2.7.0
chromadb==0.4.24
beautifulsoup4==4.12.3

# Orchestration dependencies
proxmoxer==2.0.1
paramiko==3.4.0
pyyaml==6.0.1
ansible==9.5.1

# Agent dependencies
openai==1.12.0
transformers==4.40.0

# Crypto and Financial dependencies
web3==6.15.1
aiohttp==3.9.3
ccxt==4.2.25

# Additional utilities
python-dotenv==1.0.1